src-tauri/sidecar/requirements.txt
faster-whisper>=1.0.0
piper-tts>=1.2.0
sounddevice>=0.4.6
numpy>=1.24.0
